requirements.txt
langchain
langchain_openai
langchain_community
openai
httpx==0.27.0
python-dotenv
fastapi
uvicorn
gunicorn
pypdf
geopy
